# HW/server/config/swagger.yaml
openapi: 3.0.0
# Added by API Auto Mocking Plugin
servers:
  - description: SwaggerHub API Auto Mocking
    url: ''
info:
  description: This is a simple API
  version: "1.0.0"
  title: Simple Inventory API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
tags:
  - name: Articles
    description: Some list
paths:
  /api/articles:
    get:
      tags:
        - Articles
      summary: searches articles
      operationId: searchArticles
      description: search articles list
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        '400':
          description: bad input parameter
  /api/articles/{id}:
    get:
      tags:
        - Articles
      summary: searches article by id
      operationId: searchArticleByID
      parameters:
       - name: id
         in: path
         description: ID of article to return
         required: true
         schema:
            type: integer
            format: int64
      description: search article by ID
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        '400':
          description: bad input parameter
  /api/article:
    post:
      tags:
        - Articles
      summary: add new article
      operationId: addArticle
      description: Add new article
      responses:
        '201':
          description: item created
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        description: Article to add
  /api/article/{id}:
    put:
      tags:
        - Articles
      summary: update article by ID
      operationId: updateArticleByID
      description: update article
      responses:
        '204':
          description: item updated
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        description: Article to update
  /api/list{id}:
    delete:
      tags:
        - Articles
      summary: delete article by ID
      operationId: deleteArticleByID
      description: delete article
      responses:
        '201':
          description: item deleted
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        description: Article to delete
  /api/list:
    delete:
      tags:
        - Articles
      summary: delete all articles
      operationId: deleteArticles
      description: delete articles
      responses:
        '201':
          description: items deleted
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        description: Articles to delete
components:
  schemas:
    Article:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        name:
          type: string
          example: Widget Adapter